feat(location): separate network locations
This reverts a previous introduced change to show all network locations as one.

Instead, SCELL, MCELL, and WIFI are shown separately.

This is because it is more helpful for FAEs and Sales who want to explain the differences between the different location types, and they are also sold separately (the nRF Cloud Location services behind it).